refactor: chatbot 컴포넌트 파일을 components 아래로 이동
